requirements.txt
streamlit==1.49.1
torch==2.8.0
torchvision==0.23.0
pillow==11.3.0
python-dotenv==1.1.1
groq==0.31.1